fix(expr): GET_BYTE declares integer, and a literal is named the way PG names it
Two declaration divergences the round-1 review measured, both new with this
arc and both in the sentence a client reads.

P1: GET_BYTE declared OID 20 on the wire where PostgreSQL 17.11 declares
integer (23) — `pg_typeof(get_byte('hi'::bytea,0))` is integer, measured —
beside three siblings that do declare the server's type, and pinned by the
arc's own wire cell while that cell's header comment said the other answer.
Nothing prevented RetInt32: a byte is 0..255, which is `length`'s own
declaration for the same reason. The cell moves to 23 and the binary-format
decoder learns int4.

N7: every bare integer literal was named `bigint` in the 42883 message —
this engine's own widening — where the server names it by MAGNITUDE:
integer if it fits int4, bigint if it fits int8, numeric past that and for
anything with a point or an exponent (measured: `upper(2147483647)` says
integer, `upper(2147483648)` bigint, `upper(9223372036854775808)` and
`upper(1e3)` numeric). Only the NAME changes; the domain decision is
unchanged, because every one of those is a number and no text position
takes one.

A UNARY SIGN over a literal is still that literal. `upper(-1)` is
`function upper(integer) does not exist` on the server and ANSWERED here,
because the sign hid the literal from the check — the same unwrap
isConstNumericOperand already makes.

The census asserts the MESSAGE now, per cell, where the server's sentence
is one this engine reproduces exactly. Two families deliberately differ and
are named in the struct's comment rather than asserted: PostgreSQL reports
TRIM through `pg_catalog.btrim`, and it resolves an `unknown` argument to
its closest candidate's type where this engine has no candidate list.

Reverting the get_byte declaration fails pgwire's OID cell in both result
formats; reverting numericLitDeclaredType fails the six literal-name cells
of wadjet.TestACallResolvesByItsArgumentsOrIs42883.

+++ internal/engine/expr/expr_bytea_bridge.go
@@ -82,6 +82,11 @@ func fnDecode(args []any) any {
 // value — `2202E index 5 out of valid range, 0..1` on the server, which is a
 // different class from 22P02 and names the bound. Answering NULL for it would
 // make a scan past the end look like a NULL byte.
+//
+// The result is an INT32, because `pg_typeof(get_byte('hi'::bytea,0))` is
+// `integer` on 17.11 and the wire declares what Ret says: RetInt64 put OID 20
+// on a column the server declares 23, beside three siblings that do declare
+// PostgreSQL's type (round-1 review, P1). A byte is 0..255 and always fits.
 func fnGetByte(args []any) any {
 	if len(args) < 2 || args[0] == nil || args[1] == nil {
 		return nil
@@ -91,7 +96,7 @@ func fnGetByte(args []any) any {
 	if i < 0 || i >= int64(len(raw)) {
 		raiseByteaIndexOutOfRange(i, len(raw))
 	}
-	return int64(raw[i])
+	return int32(raw[i])
 }
 
 // fnSetByte writes ONE byte and answers the new value. The byte is taken mod
